fix: много исправлений userController

+++ src/main/java/ru/yandex/practicum/filmorate/controller/FilmController.java
@@ -5,9 +5,11 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 import ru.yandex.practicum.filmorate.exceptions.FilmNotFoundException;
 import ru.yandex.practicum.filmorate.exceptions.FilmValidationException;
+import ru.yandex.practicum.filmorate.exceptions.UserNotFoundException;
 import ru.yandex.practicum.filmorate.model.Film;
 import ru.yandex.practicum.filmorate.service.FilmService;
 import ru.yandex.practicum.filmorate.storage.InMemoryFilmStorage;
+import ru.yandex.practicum.filmorate.storage.InMemoryUserStorage;
 import ru.yandex.practicum.filmorate.validators.FilmValidator;
 
 import java.util.Collection;
@@ -20,17 +22,19 @@ public class FilmController {
 
     private final InMemoryFilmStorage filmStorage;
     private final FilmService filmService;
+    private final InMemoryUserStorage userStorage;
 
     @Autowired
-    public FilmController(InMemoryFilmStorage filmStorage, FilmService filmService) {
+    public FilmController(InMemoryFilmStorage filmStorage, FilmService filmService, InMemoryUserStorage userStorage) {
         this.filmStorage = filmStorage;
         this.filmService = filmService;
+        this.userStorage = userStorage;
     }
 
 
     @GetMapping("/films")
     public Collection<Film> getFilms() {
-        return filmStorage.getFilms();
+        return filmStorage.getFilms().values();
     }
 
     @PostMapping("/films")
@@ -44,33 +48,33 @@ public class FilmController {
     }
 
     @PutMapping("/films")
-    public Optional<Film> putFilm(@RequestBody Optional<Film> film) {
-        if (film.isEmpty()) {
-            throw new FilmNotFoundException("Фильм не найден!");
+    public Optional<Film> putFilm(@RequestBody Film film) {
+        if (!FilmValidator.validate(film)) {
+            log.error("Ошибка валидации фильма при запросе PUT /users");
+            throw new FilmValidationException("Ошибка валидации фильма. Проверьте данные.");
         }
-        if (!FilmValidator.validate(film.get())) {
-            log.error("Ошибка валидации фильма при запросе PUT /films");
-            throw new FilmValidationException("Ошибка валидации фильма. Проверьте данные!");
+        if (!filmStorage.containsFilm(film.getId())) {
+            throw new FilmNotFoundException("Такой фильм не найден!");
+        } else {
+            filmStorage.getFilms().replace(film.getId(), film);
+            return Optional.of(filmStorage.getFilms().get(film.getId()));
         }
-        try {
-            filmStorage.updateFilm(film.get());
-            return Optional.of(filmStorage.getFilm(film.get().getId()));
-        } catch (FilmNotFoundException ex) {
-            System.out.println(ex.getMessage());
-            log.error("Ошибка обновления фильма при запросе PUT /films");
-            System.out.println("Ошибка при обновлении данных фильма");
-        }
-        return Optional.empty();
     }
 
     @PutMapping("films/{id}/like/{userId}")
-    public void addLikeToFilm(@PathVariable Integer id, @PathVariable Integer userId) {
+    public void addLikeToFilm(@PathVariable Integer id, @PathVariable Long userId) {
         filmService.addLikeToFilm(userId, id);
     }
 
     @DeleteMapping("films/{id}/like/{userId}")
-    public void removeLikeFromFilm(@PathVariable Integer id, @PathVariable Integer userId) {
-        filmService.removeLikeToFilm(id, userId);
+    public void removeLikeFromFilm(@PathVariable Integer id, @PathVariable Long userId) {
+        if (!filmStorage.containsFilm(id)) {
+            throw new FilmNotFoundException("Фильм с таким ID не найден!");
+        } else if (!userStorage.containsId(userId)) {
+            throw new UserNotFoundException("Пользователь с таким ID не найден!");
+        } else {
+            filmService.removeLikeToFilm(id, userId);
+        }
     }
 
     @GetMapping("/films/popular")
@@ -78,7 +82,7 @@ public class FilmController {
         if (count.isEmpty()) {
             return filmService.getMostPopularFilms(10);
         } else if (count.get() > filmStorage.getFilms().size()) {
-            throw new FilmNotFoundException("Некорректный запрос, такого кол-ва фильмов нет!");
+            return filmService.getMostPopularFilms(filmStorage.getFilms().size());
         } else {
             return filmService.getMostPopularFilms(count.get());
         }
